MovedFieldInfoExtensions into its own file
